requirements.txt
newspaper4k==0.9.2
aiohttp==3.9.3
pandas==2.2.0